style(init): Modularize entire init.lua
Now each plugin is in it's own seperate file that returns the plugin and
it's configs. This makes it simpler for me apply lazy loading to this
config.

Settings, Keymaps, Yank-highlight are inside kickstart directory within
their seperate files.
