+++ lua/kickstart/plugins/comment.lua
@@ -1,0 +1,5 @@
+return {
+  -- "gc" to comment visual regions/lines
+  'numToStr/Comment.nvim',
+  opts = {}
+}
