+++ lua/kickstart/plugins/git-related.lua
@@ -1,0 +1,6 @@
+-- NOTE: First, some plugins that don't require any configuration
+return {
+  -- Git related plugins
+  'tpope/vim-fugitive',
+  'tpope/vim-rhubarb',
+}
