+++ lua/kickstart/plugins/indent-blankline.lua
@@ -1,0 +1,8 @@
+return {
+  -- Add indentation guides even on blank lines
+  'lukas-reineke/indent-blankline.nvim',
+  -- Enable `lukas-reineke/indent-blankline.nvim`
+  -- See `:help ibl`
+  main = 'ibl',
+  opts = {},
+}
