+++ lua/kickstart/plugins/vim-seluth.lua
@@ -1,0 +1,4 @@
+return {
+  -- Detect tabstop and shiftwidth automatically
+  'tpope/vim-sleuth',
+}
